Tests ok + decrease dependencies number in Controllers + start optimizing fixtures
